Extract actionFunction into an own JS module
Several hooks will use this function. Therefore move it to an own JS
module.
